backend/requirements.txt
# ⚠️ DEPRECATED: 이 파일은 더 이상 Docker 빌드에 사용되지 않습니다.
# 의존성 기준: 루트 pyproject.toml + uv.lock
# 로컬 개발 시에도 `uv sync` 사용을 권장합니다.
#
Django>=5.0,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
django-q2>=1.6.0
django-cors-headers>=4.0
django-filter>=23.0
djangorestframework-simplejwt>=5.3
django-allauth>=0.61.1
dj-rest-auth>=5.0.0
requests>=2.31.0
cryptography>=41.0.0
django-axes>=6.0.0
Pillow>=12.1.1
gunicorn>=22.0.0
